+++ components/inplaceeditor/src/inplace-editor/inplaceeditor.component.ts
@@ -1,10 +1,11 @@
 import Vue from 'vue';
+import { isUndefined } from '@syncfusion/ej2-base';
 import { ComponentBase, EJComponentDecorator } from '@syncfusion/ej2-vue-base';
 import { InPlaceEditor } from '@syncfusion/ej2-inplace-editor';
 
 
 export const properties: string[] = ['actionOnBlur', 'adaptor', 'cancelButton', 'cssClass', 'disabled', 'editableOn', 'emptyText', 'enableEditMode', 'enableHtmlSanitizer', 'enablePersistence', 'enableRtl', 'locale', 'mode', 'model', 'name', 'popupSettings', 'primaryKey', 'saveButton', 'showButtons', 'submitOnEnter', 'template', 'textOption', 'type', 'url', 'validationRules', 'value', 'actionBegin', 'actionFailure', 'actionSuccess', 'beforeSanitizeHtml', 'beginEdit', 'cancelClick', 'change', 'created', 'destroyed', 'submitClick', 'validating'];
-export const modelProps: string[] = [];
+export const modelProps: string[] = ['value'];
 
 /**
  * `ejs-inplace-editor` represents the VueJS InPlaceEditor Component.
@@ -13,7 +14,10 @@ export const modelProps: string[] = [];
  * ```
  */
 @EJComponentDecorator({
-    props: properties
+    props: properties,
+    model: {
+        event: 'modelchanged'
+    }
 })
 export class InPlaceEditorComponent extends ComponentBase {
     
@@ -27,7 +31,9 @@ export class InPlaceEditorComponent extends ComponentBase {
     
     constructor() {
         super();
-        this.ej2Instances = new InPlaceEditor({});
+        this.ej2Instances = new InPlaceEditor({});        this.ej2Instances._trigger = this.ej2Instances.trigger;
+        this.ej2Instances.trigger = this.trigger;
+
         this.bindProperties();
         this.ej2Instances._setProperties = this.ej2Instances.setProperties;
         this.ej2Instances.setProperties = this.setProperties;
@@ -45,6 +51,26 @@ export class InPlaceEditorComponent extends ComponentBase {
                 });
             });
         }
+    }
+    public trigger(eventName: string, eventProp: {[key:string]:Object}, successHandler?: Function): void {
+        if ((eventName === 'change' || eventName === 'input') && this.models && (this.models.length !== 0)) {
+            let key: string[] = this.models.toString().match(/checked|value/) || [];
+            let propKey: string = key[0];
+            if (eventProp && key && !isUndefined(eventProp[propKey])) {
+                (this as any).$emit('update:'+ propKey, eventProp[propKey]);
+                (this as any).$emit('modelchanged', eventProp[propKey]);
+            }
+        } else if ((eventName === 'actionBegin' && eventProp.requestType === 'dateNavigate') && this.models && (this.models.length !== 0)) {
+            let key: string[] = this.models.toString().match(/currentView|selectedDate/) || [];
+            let propKey: string = key[0];
+            if (eventProp && key && !isUndefined(eventProp[propKey])) {
+                (this as any).$emit('update:'+ propKey, eventProp[propKey]);
+                (this as any).$emit('modelchanged', eventProp[propKey]);
+            }
+        }
+        if (this.ej2Instances && this.ej2Instances._trigger) {
+            this.ej2Instances._trigger(eventName, eventProp, successHandler);
+        }            
     }
 
     public render(createElement: any) {