+++ components/navigations/src/tab/tab.component.ts
@@ -4,7 +4,7 @@ import { Tab } from '@syncfusion/ej2-navigations';
 import { TabItemsDirective, TabItemDirective, TabItemsPlugin, TabItemPlugin } from './items.directive'
 
 
-export const properties: string[] = ['animation', 'cssClass', 'enableHtmlSanitizer', 'enablePersistence', 'enableRtl', 'headerPlacement', 'height', 'heightAdjustMode', 'items', 'loadOn', 'locale', 'overflowMode', 'scrollStep', 'selectedItem', 'showCloseButton', 'width', 'added', 'adding', 'created', 'destroyed', 'removed', 'removing', 'selected', 'selecting'];
+export const properties: string[] = ['allowDragAndDrop', 'animation', 'cssClass', 'dragArea', 'enableHtmlSanitizer', 'enablePersistence', 'enableRtl', 'headerPlacement', 'height', 'heightAdjustMode', 'items', 'loadOn', 'locale', 'overflowMode', 'scrollStep', 'selectedItem', 'showCloseButton', 'width', 'added', 'adding', 'created', 'destroyed', 'dragged', 'dragging', 'onDragStart', 'removed', 'removing', 'selected', 'selecting'];
 export const modelProps: string[] = [];
 
 /**
@@ -62,6 +62,10 @@ export class TabComponent extends ComponentBase {
 
     public enableTab(index: number, value: boolean): void {
         return this.ej2Instances.enableTab(index, value);
+    }
+
+    public getItemIndex(tabItemId: string): number {
+        return this.ej2Instances.getItemIndex(tabItemId);
     }
 
     public hideTab(index: number, value?: boolean): void {